oracle-scripts/stock-daily-analyzer/requirements.txt
pykrx>=1.0.45
gspread>=6.0.0
google-auth>=2.25.0
anthropic>=0.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
